// src/lib/api/generated/model/appointmentStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Agendly.Api
 * OpenAPI spec version: 1.0
 */

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AppointmentStatus = {
  Scheduled: 'Scheduled',
  Confirmed: 'Confirmed',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
  NoShow: 'NoShow',
} as const;